pyproject.toml
[tool.poetry]
name = "nnp-university"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
envparse = "^0.2.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.8"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
